+++ Contents/mods/KR CoreOS/media/lua/server/KRCore_Server.lua
@@ -1,5 +1,5 @@
 -- ==========================================================================
--- KR CoreOS - Server v1.1.0 B41.78.7 (Server)
+-- KR CoreOS - Server v1.2.0 B41.78.7 (Server)
 -- Copyright (C) 2026 D4RK-C0MP4N1. Licensed under the MIT License (see LICENSE).
 -- ==========================================================================
 --
@@ -180,7 +180,7 @@ end
 --   modify it.
 
 local function onPostDistributionMerge()
-    print(TAG .. "Starting KR CoreOS v1.1.0 (B41)...")
+    print(TAG .. "Starting KR CoreOS v1.2.0 (B41)...")
     processContentQueue()
     processDistributions()
     KRCore._initialized = true